feature: controller 작성, 잘못된 폴더 구조 수정
